requirements.txt
flask==3.0.0
tensorflow==2.13.0
pillow==9.5.0
numpy==1.24.3
werkzeug==3.0.1
